+++ app/src/main/java/nextstep/jwp/exception/NoMatchingControllerException.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.controller;
+package nextstep.jwp.exception;
 
 public class NoMatchingControllerException extends RuntimeException {
 }
