+++ app/src/main/java/nextstep/jwp/exception/NoSuchContentTypeException.java
@@ -1,0 +1,4 @@
+package nextstep.jwp.exception;
+
+public class NoSuchContentTypeException extends RuntimeException {
+}
