+++ app/src/main/java/nextstep/jwp/exception/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http;
+package nextstep.jwp.exception;
 
 public class ResourceNotFoundException extends RuntimeException {
 }
