--- app/src/main/java/nextstep/jwp/http/NoSuchContentException.java
@@ -1,4 +1,0 @@
-package nextstep.jwp.http;
-
-public class NoSuchContentException extends RuntimeException {
-}
